+++ src/world/objects/Surface.cpp
@@ -46,6 +46,7 @@ Surface::applyTransform(std::shared_ptr<render::Primitive> primitive) const {
   auto result = std::make_shared<render::Instance>(primitive);
   result->setMatrix(localTransform());
   result->setVelocity(m_velocity);
+  result->setRenderTextureSubview(primitive->renderTextureSubview());
   return result;
 }
 
@@ -67,6 +68,7 @@ std::shared_ptr<render::Primitive> Surface::toRaytracer(render::Scene* scene,
   if (!primitive) {
     return primitive;
   }
+  primitive->setRenderTextureSubview(m_renderTextureSubview.toStdString());
 
   if (material()) {
     primitive->setMaterial(material()->toRaytracerMaterial());